split routers class
