+++ model/db_connection.py
@@ -1,0 +1,22 @@
+import mysql.connector
+import mysql.connector.pooling
+import logging
+import os
+
+def connection():
+  try:
+    dbconfig = {
+        "host": os.getenv("DBHOST"),
+        "user": os.getenv("DBUSER"),
+        "password": os.getenv("DBPASSWORD"),
+        "database":"wehelp_stage2_taipei_spot",
+    }
+    cnxpool = mysql.connector.pooling.MySQLConnectionPool(
+      pool_name="mypool",
+      pool_size=3,
+      **dbconfig
+    )
+    cnx1 = cnxpool.get_connection()
+    return cnx1
+  except:
+    logging.warning("database connection fail")
